+++ lib/src/style/style.dart
@@ -131,6 +131,26 @@ class TileCacheBudget {
   TileCacheBudget(this.type, this.size);
 }
 
+/// The description of the raster data layers and the bands contained within the tiles.
+@experimental
+class RasterDataLayer {
+  /// Identifier of the data layer fetched from tiles.
+  String layerId;
+
+  /// An array of bands found in the data layer.
+  List<String> bands;
+
+  Map<String, List<String>> toJson() => {layerId: bands};
+
+  static RasterDataLayer? decode(Object? layer) {
+    var layerObject = Map<String, dynamic>.from(layer as Map<dynamic, dynamic>)
+        .cast<String, dynamic>();
+    return RasterDataLayer(layerObject.keys.first, layerObject.values.first);
+  }
+
+  RasterDataLayer(this.layerId, this.bands);
+}
+
 /// Define the duration and delay for a style transition.
 class StyleTransition {
   StyleTransition({this.duration, this.delay});
@@ -313,6 +333,9 @@ extension StyleSource on StyleManager {
         break;
       case "raster":
         source = RasterSource(id: sourceId);
+        break;
+      case "raster-array":
+        source = RasterArraySource(id: sourceId);
         break;
       default:
         print("Source type: $type unknown.");